move listeners to new loc
